requirements.txt
psycopg2-binary
python-dotenv
openai
requests
tqdm
fastapi
uvicorn
python-multipart
requests-toolbelt
